feat: introduce context-sensitive logs
breaking: Sync will now fail when encountering unresolvable page mentions.
Before this change, it was difficult to pinpoint where an unresolvable page
mention came from. Rendering thus included unresolved mentions as an annotation
in the rendered output. This could lead to incomplete text, e.g. if an
important mention was suddenly missing from a sentence.

With the context-sensitive logs outputting the page that produced an error,
and an URL to the mentioned page, users should now have enough info
from log outputs alone to locate the error.