feat: render links as relative paths
We previously rendered all page links as absolute paths e.g.
`/page.md`. This created some problems
- some page generators that use base path prefixes like `/docs/page.md`
  and thus need to rewrite these links, which is brittle
- text editors like VSCode cannot follow the markdown links properly
  because they think the links are absolute to the root `/` of the fs
- tools like textlint-rule-no-dead-link expect relative links to resolve
  them correctly

Now rendering them as relative paths.

As part of these changes it was also necessary to switch properties
parsing to render properties only as plaintext, not markdown. This is
fine because we consider frontmatter to be data (YML), not markup. Most
site generators cannot deal with markdown in frontmatter anyway.
This change was necessary because we need to parse page properties
before we can determine the destination path of the markdown file where
the page will be rendered to, and that would now require resolving
relative links.